Update Scribe.py
Bug Fix: Can't close subsequent scribe tabs after Preference tab is added.

There was an issue where the input fields array was out of alignment with the tab list array after adding a preference tab because the preference tab didn't have a text field. 
Fix: Generated a dummy field in memory when adding the tab to the screen and add the dummy text field to the list of text fields. When removing the preference tab, remove the dummy field with pop()