Update Scribe.py
This updates the close_tab logic in the event the tab name is blank but there is still content to save. 
This was intended to introduce an interrupter, but that isn't the most intuitive way to save that file in a hurry. 
This way, the filename would be named with the current date/hr of that save. So, it will naturally overwrite anything else in that hour as well. This helps with keeping files organized and without duplicates